First working version of Uthernet I as a Card.
