+++ source/Configuration/PageConfigTfe.h
@@ -2,6 +2,7 @@
 
 #include "IPropertySheetPage.h"
 #include "../Tfe/Uilib.h"
+#include "../Card.h"
 
 #include <string>
 
@@ -11,13 +12,13 @@ public:
 	CPageConfigTfe()
 	{
 		CPageConfigTfe::ms_this = this;
-		m_tfe_enabled = 0;
+		m_tfe_enabled = CT_Empty;
 	}
 	virtual ~CPageConfigTfe(){}
 
 	static INT_PTR CALLBACK DlgProc(HWND window, UINT message, WPARAM wparam, LPARAM lparam);
 
-	int m_tfe_enabled;
+	SS_CARDTYPE m_tfe_enabled;
 	std::string m_tfe_interface_name;
 
 protected:
