+++ source/Configuration/PropertySheetHelper.cpp
@@ -32,6 +32,7 @@ Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307  USA
 #include "../Log.h"
 #include "../Registry.h"
 #include "../SaveState.h"
+#include "../Tfe/tfe.h"
 
 /*
 Config causing AfterClose msgs:
